% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loader.R
\name{loader}
\alias{loader}
\title{A custom UFOKN themed loading gif}
\usage{
loader(ui_element)
}
\arguments{
\item{ui_element}{A UI element that should include a loader while the
corresponding output is being calculated}
}
\description{
Add a loader that shows when an output element is calculating
}
