% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{ufokn_title}
\alias{ufokn_title}
\title{Add title to Shiny App with UF-OKN Logo}
\usage{
ufokn_title(title)
}
\arguments{
\item{title}{Title of shiny app}
}
\description{
Add title to Shiny App with UF-OKN Logo
}
