% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown.R
\name{use_pkgdown}
\alias{use_pkgdown}
\title{Use UF-OKN themed pkgdown site}
\usage{
use_pkgdown(overwrite = FALSE)
}
\arguments{
\item{overwrite}{logical; should an existing _pkgdown.yml file be overwritten?}
}
\description{
Use UF-OKN themed pkgdown site
}
